@@ -1,7 +1,7 @@
 {
   "name": "plaid",
   "productName": "Plaid",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "author": {
     "name": "Piotr Marczyński",
     "email": "piotmski@gmail.com"
@@ -23,7 +23,7 @@
       "build"
     ],
     "win": {
-      "icon": "build/assets/icon/icon.ico",
+      "icon": "build/assets/appicon/icon.ico",
       "extraFiles": [
         {
           "from": "src/assets/Plaid.VisualElementsManifest.xml",
@@ -35,7 +35,7 @@
       ]
     },
     "linux": {
-      "icon": "assets/icon/icon.png"
+      "icon": "assets/appicon/icon.png"
     }
   },
   "scripts": {
@@ -50,34 +50,34 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@openfonts/comfortaa_latin": "^1.44.2",
+    "@angular/animations": "^9.0.2",
+    "@angular/common": "^9.0.2",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
+    "@angular/forms": "^9.0.2",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/router": "^9.0.2",
+    "@openfonts/comfortaa_latin": "^1.44.3",
     "@openfonts/source-sans-pro_all": "^1.44.0",
-    "electron-updater": "^4.2.0",
-    "mousetrap": "^1.6.3",
+    "electron-updater": "^4.2.2",
+    "mousetrap": "^1.6.5",
     "ngx-electron": "~2.2.0",
     "rxjs": "~6.5.4",
-    "tslib": "~1.10.0",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.11.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.22",
-    "@angular/cli": "~8.3.22",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/node": "~12.12.24",
-    "codelyzer": "^5.2.1",
-    "electron": "^7.1.9",
-    "electron-builder": "^22.2.0",
+    "@angular-devkit/build-angular": "^0.900.3",
+    "@angular/cli": "^9.0.3",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular/language-service": "^9.0.2",
+    "@types/node": "^12.12.28",
+    "codelyzer": "^5.1.2",
+    "electron": "^8.0.1",
+    "electron-builder": "^22.3.2",
     "ts-node": "~8.6.2",
     "tslint": "~5.20.1",
-    "typescript": "~3.5.3"
+    "typescript": "~3.7.5"
   }
 }
