@@ -1,4 +1,4 @@
-import {Component, EventEmitter, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {PlaidFacade} from '../../plaid.facade';
 import {AuthInfo} from '../../models/auth-info';
 import {HttpErrorResponse} from '@angular/common/http';
@@ -11,17 +11,21 @@ import {User} from '../../models/user';
   styleUrls: ['./connection-issue-resolver.component.scss']
 })
 export class ConnectionIssueResolverComponent implements OnInit {
-  loginPopupVisible = false;
-  lostConnectionPopupVisible = false;
-  errorPopupVisible = false;
+  loginModalVisible = false;
+  lostConnectionModalVisible = false;
+  errorModalVisible = false;
   authInfo: AuthInfo = { jiraUrl: null, username: null, password: null };
   _error: HttpErrorResponse;
   fetching = false;
   _currentUser: User = null;
   submittedAuthInfo: AuthInfo = null;
+  httpNoticeVisible = false;
 
   @Input()
   changeCredentials: EventEmitter<void>;
+
+  @Output()
+  modalVisible = new EventEmitter<boolean>();
 
   constructor(private facade: PlaidFacade) {}
 
@@ -30,10 +34,11 @@ export class ConnectionIssueResolverComponent implements OnInit {
     this.facade.getAuthError$().pipe(skip(1)).subscribe((authError: HttpErrorResponse) => this.error = authError);
     this.facade.getAuthenticatedUser$().pipe(skip(1)).subscribe(user => this.currentUser = user);
     this.changeCredentials.subscribe(() => {
-      this.loginPopupVisible = true;
+      this.loginModalVisible = true;
       this._error = null;
       this.fetching = false;
       this.authInfo = this.facade.getAuthInfo() ? this.facade.getAuthInfo() : { jiraUrl: null, username: null, password: null };
+      this.modalVisible.emit(true);
     });
   }
 
@@ -42,12 +47,15 @@ export class ConnectionIssueResolverComponent implements OnInit {
     this.fetching = false;
     if (!this.currentUser || error && [401, 403].indexOf(error.status) !== -1) { // Authentication error
       this.facade.discardAuthenticatedUser();
-      this.loginPopupVisible = true;
+      this.loginModalVisible = true;
       this.authInfo = this.facade.getAuthInfo() ? this.facade.getAuthInfo() : { jiraUrl: null, username: null, password: null };
+      this.modalVisible.emit(true);
     } else if (error && error.status === 0) { // Network connection issue
-      this.lostConnectionPopupVisible = true;
+      this.lostConnectionModalVisible = true;
+      this.modalVisible.emit(true);
     } else if (error) { // Unknown error
-      this.errorPopupVisible = true;
+      this.errorModalVisible = true;
+      this.modalVisible.emit(true);
     }
   }
   get error(): HttpErrorResponse {
@@ -57,23 +65,33 @@ export class ConnectionIssueResolverComponent implements OnInit {
   set currentUser(user: User) {
     this._currentUser = user;
     if (user) {
-      this.loginPopupVisible = false;
-      this.lostConnectionPopupVisible = false;
-      this.errorPopupVisible = false;
+      this.loginModalVisible = false;
+      this.lostConnectionModalVisible = false;
+      this.errorModalVisible = false;
+      this.modalVisible.emit(false);
     }
   }
   get currentUser(): User {
     return this._currentUser;
   }
 
-  stripUrl(): void {
-    if (this.authInfo.jiraUrl && this.authInfo.jiraUrl.length > 0 && this.authInfo.jiraUrl[this.authInfo.jiraUrl.length - 1] === '/') {
-      this.authInfo.jiraUrl = this.authInfo.jiraUrl.substr(0, this.authInfo.jiraUrl.length - 1);
+  onUrlChange(): void {
+    if (this.authInfo.jiraUrl && this.authInfo.jiraUrl.length > 0) {
+      // Remove trailing slash
+      if (this.authInfo.jiraUrl[this.authInfo.jiraUrl.length - 1] === '/') {
+        this.authInfo.jiraUrl = this.authInfo.jiraUrl.substr(0, this.authInfo.jiraUrl.length - 1);
+      }
+      // Add protocol if missing
+      if (this.authInfo.jiraUrl.substr(0, 7) !== 'http://' && this.authInfo.jiraUrl.substr(0, 8) !== 'https://') {
+        this.authInfo.jiraUrl = 'https://' + this.authInfo.jiraUrl;
+      }
+      // Display http notice if applicable
+      this.httpNoticeVisible = this.authInfo.jiraUrl.substr(0, 7) === 'http://';
     }
   }
 
   submit(): void {
-    this.stripUrl();
+    this.onUrlChange();
     this.submittedAuthInfo = {...this.authInfo};
     this.facade.setAuthInfo(this.submittedAuthInfo);
     this.facade.discardAuthenticatedUser();
