@@ -1,6 +1,5 @@
 import {
-  ChangeDetectionStrategy,
-  ChangeDetectorRef,
+  ChangeDetectionStrategy, ChangeDetectorRef,
   Component,
   ElementRef,
   Input,
@@ -10,7 +9,7 @@ import {
 } from '@angular/core';
 import {Worklog} from '../../models/worklog';
 import {PlaidFacade} from '../../plaid.facade';
-import {Subscription} from 'rxjs';
+import {WorklogPanelsManagerService} from './worklog-panels-manager.service';
 
 @Component({
   selector: 'plaid-worklog-panel',
@@ -24,81 +23,72 @@ export class WorklogPanelComponent implements OnInit, OnDestroy {
   _pixelsPerMinute: number;
   undersized = false;
   tooLow = false;
-  subscriptions: Subscription[] = [];
   viewDestroyed = false;
+  date: Date;
+  panelWidth: number;
+  panelHeight: number;
+  maxHeight: number;
+  panelOffsetTop: number;
+  panelOffsetLeft: number;
+  panelHue: number;
+  components: string;
+  timeRange: string;
+  _darkMode: boolean;
 
   @Input()
   set worklog(worklog: Worklog) {
     this._worklog = worklog;
-    if (this.pixelsPerMinute) {
-      setTimeout(() => this.checkSizeAndPosition());
-    }
+    this.date = new Date(this.worklog.started);
+    this.panelWidth = 1 / this.worklog._columns;
+    this.panelOffsetLeft = this.worklog._column * this.panelWidth;
+    this.panelHue = Math.round((Number(this.worklog.issue.fields.parent
+      ? this.worklog.issue.fields.parent.id
+      : this.worklog.issue.id) * 360 / 1.61803)) % 360;
+    this.components = this.worklog.issue.fields.components
+      ? this.worklog.issue.fields.components.map(c => c.name).join(', ')
+      : null;
+    this.computeHeightAndOffset();
+    this.computeTimeRange();
+    this.manager.scheduleCheckSizeAndPosition();
   }
   get worklog(): Worklog {
     return this._worklog;
   }
+
   @Input()
   set pixelsPerMinute(pixelsPerMinute: number) {
     this._pixelsPerMinute = pixelsPerMinute;
-    if (this.worklog) {
-      setTimeout(() => this.checkSizeAndPosition());
-    }
+    this.computeHeightAndOffset();
+    this.manager.scheduleCheckSizeAndPosition();
   }
   get pixelsPerMinute(): number {
     return this._pixelsPerMinute;
   }
 
+  set darkMode(value: boolean) {
+    this._darkMode = value;
+    this.cdr.markForCheck();
+  }
+  get darkMode(): boolean {
+    return this._darkMode;
+  }
+
   @ViewChild('panelInner', { static: true })
   panelInner: ElementRef;
 
-  constructor(private facade: PlaidFacade, private cdr: ChangeDetectorRef) { }
+  constructor(private facade: PlaidFacade, private cdr: ChangeDetectorRef, private manager: WorklogPanelsManagerService) { }
 
   ngOnInit(): void {
-    this.subscriptions.push(this.facade.getJiraURL$().subscribe(url => this.jiraURL = url));
-    this.subscriptions.push(this.facade.windowResize$().subscribe(() => this.checkSizeAndPosition()));
+    this.jiraURL = this.facade.getJiraURL();
+    this.manager.addPanel(this);
   }
 
   ngOnDestroy(): void {
+    this.manager.removePanel(this);
     this.viewDestroyed = true;
-    while (this.subscriptions.length > 0) {
-      this.subscriptions.pop().unsubscribe();
-    }
   }
 
-  get date(): Date {
-    return new Date(this.worklog.started);
-  }
-
-  get panelWidth(): number {
-    return 1 / this.worklog._columns;
-  }
-
-  get panelHeight(): number {
-    return Math.min(this.worklog.timeSpentSeconds / 60 * this.pixelsPerMinute, this.maxHeight);
-  }
-
-  get maxHeight(): number {
-    return 1440 * this.pixelsPerMinute - this.panelOffsetTop;
-  }
-
-  get panelOffsetTop(): number {
-    return (this.date.getHours() * 60 + this.date.getMinutes()) * this.pixelsPerMinute;
-  }
-
-  get panelOffsetLeft(): number {
-    return this.worklog._column * this.panelWidth;
-  }
-
-  get panelHue(): number {
-    const num = Number(this.worklog.issue.fields.parent ? this.worklog.issue.fields.parent.id : this.worklog.issue.id);
-    return Math.round((num * 360 / 1.61803)) % 360;
-  }
-
-  get components(): string {
-    return this.worklog.issue.fields.components ? this.worklog.issue.fields.components.map(c => c.name).join(', ') : null;
-  }
-
-  get timeRange(): string {
+  computeTimeRange(): void {
     const startTime = new Date(this.worklog.started);
     const endTime = new Date(startTime);
     endTime.setTime(endTime.getTime() + this.worklog.timeSpentSeconds * 1000);
@@ -108,7 +98,7 @@ export class WorklogPanelComponent implements OnInit, OnDestroy {
       startTime.getMonth() === endTime.getMonth() &&
       startTime.getDate() === endTime.getDate()
     ) {
-      return startTime.toLocaleTimeString(undefined, { hour: 'numeric', minute: '2-digit' }) + ' - ' +
+      this.timeRange = startTime.toLocaleTimeString(undefined, { hour: 'numeric', minute: '2-digit' }) + ' - ' +
         endTime.toLocaleTimeString(undefined, { hour: 'numeric', minute: '2-digit' });
     } else {
       let sumOfSeconds = this.worklog.timeSpentSeconds;
@@ -117,16 +107,22 @@ export class WorklogPanelComponent implements OnInit, OnDestroy {
       const minutes: number = Math.floor(sumOfSeconds / 60);
       sumOfSeconds -= minutes * 60;
       const seconds: number = sumOfSeconds;
-      return 'Since ' + startTime.toLocaleTimeString(undefined, { hour: 'numeric', minute: '2-digit' }) +
+      this.timeRange = 'Since ' + startTime.toLocaleTimeString(undefined, { hour: 'numeric', minute: '2-digit' }) +
         ' for ' + (hours ? hours + 'h ' : '') + (minutes ? minutes + 'm ' : '') + (seconds ? seconds + 's' : '');
     }
+  }
+
+  computeHeightAndOffset(): void {
+    this.panelOffsetTop = (this.date.getHours() * 60 + this.date.getMinutes()) * this.pixelsPerMinute;
+    this.maxHeight = 1440 * this.pixelsPerMinute - this.panelOffsetTop;
+    this.panelHeight = Math.min(this.worklog.timeSpentSeconds / 60 * this.pixelsPerMinute, this.maxHeight);
   }
 
   checkSizeAndPosition(): void {
     if (!this.viewDestroyed) {
       this.undersized = this.panelInner.nativeElement.scrollHeight > this.panelHeight;
       this.tooLow = this.panelInner.nativeElement.scrollHeight + 1 > this.maxHeight;
-      this.cdr.detectChanges();
+      this.cdr.markForCheck();
     }
   }
 }
