@@ -13,6 +13,8 @@ import {RefreshButtonComponent} from './components/refresh-buton/refresh-button.
 import {ExtHrefDirective} from './directives/ext-href.directive';
 import {NgxElectronModule} from 'ngx-electron';
 import {AuthStatusComponent} from './components/auth-status/auth-status.component';
+import {ZoomControlsComponent} from './components/zoom-controls/zoom-controls.component';
+import {CurrentTimeMarkerComponent} from './components/current-time-marker/current-time-marker.component';
 
 @NgModule({
   declarations: [
@@ -24,7 +26,9 @@ import {AuthStatusComponent} from './components/auth-status/auth-status.componen
     PlannerGridBackgroundComponent,
     RefreshButtonComponent,
     ExtHrefDirective,
-    AuthStatusComponent
+    AuthStatusComponent,
+    ZoomControlsComponent,
+    CurrentTimeMarkerComponent
   ],
   imports: [
     BrowserModule,
